+++ disk/DiskFile.cpp
@@ -771,6 +771,7 @@ void XRootdDirectory::rmdir() {
 bool XRootdDirectory::exist() {
   XrdCl::StatInfo *statInfo;
   XrdCl::XRootDStatus statStatus = m_xrootFileSystem.Stat(m_truncatedDirectoryURL,statInfo,c_xrootTimeout);
+  cta::exception::XrootCl::throwOnError(statStatus,"In XRootdDirectory::exist() : failed to stat the directory at "+m_URL);
   if(statStatus.errNo == XErrorCode::kXR_NotFound){
     return false;
   }